updated file name from night read to translator
