--- lib/night_write.rb
@@ -1,5 +1,0 @@
-require './lib/night_writer'
-
-input = File.open(ARGV[0], 'r')
-contents = input.read
-input.close
